Fix regression caused by dropping the last point
We started dropping the last point in
ceb658c0a764232cfacae30085eb0af423deb404

This caused a regression because sometimes the last point is from the current
trip, just with speed and distance = 0. Add a new check for the time delta in
addition to speed and distance.

This fixes the regressions