Add ipv6 routes to both security group and routing tables
Although people won't see the ipv6 until they start to use it.  Note that there
are a bunch of manual steps to turn on IPv6 for this setup.  This change merely
automates the tedious work of setting up the routing tables and security
groups.
https://github.com/e-mission/e-mission-server/issues/530#issuecomment-354061649

At this point, I declare that I am done with tweaking the configuration and
will use the configuration deployed from this template (including
75d19de346a04cb51142e048e9eea05b1d15b44c,
7a32bb69f232f312cf97c4089d8edbd14be642eb...) as the setup for the
default/reference e-mission server.