Handle the case in which the user has only one place
We had already handled the case in which the user had no places.
This also handles the case in which the user has exactly one place.
We ran into this while resetting the pipeline while re-running after moving to
the more scalable server.

In that case too, we return None and skip the reset.