Add support for converting to sandbox tokens on firebase
Although firebase does not support sandbox v/s production for the send API,
they do support it for the import/conversion API. And if you use
`sandbox=False` while trying to convert a token from a debug build, the send
will fail with "Not registered" or "Invalid Token".

Since we already have a dev flag, let's hook it up to this.
https://github.com/e-mission/e-mission-server/issues/564#issuecomment-360725717
https://github.com/e-mission/e-mission-server/issues/564#issuecomment-360728695
and make it easier to test