Add a simple test that just copies files over
Tests are failing on jenkins although they pass locally because the seed model
is not found. But I'm copying the seed model!

Let's add a very simple test that basically just copies over the seed model so
that we can see the failure (if any) in the console.

Also switch the default test run over to this simple test