Couple of fixes found while testing after removing the inference results
Note that these can also happen if there is a request between the time that the
data is cleaned and the mode is inferred. We need to be more robust to these
changes anyway.
- if there is no inference result yet, don't override the mode
- if there is no inference result yet, return the cleaned_resampled timestamp
  as the completed timestamp

Both of these will fall back to the cleaned data if the inferred data is not
present