Add support for loading entries directly into the usercache
If the intake pipeline is not running on the server
https://github.com/e-mission/e-mission-docs/blob/master/docs/e-mission-server/deploying_your_own_server_to_production.md#the-analysis-pipeline
then data will stay in the usercache forever.

So it won't be passed through the formatter, and will not have `data.fmt_time`
and `metadata.write_fmt_time` filled in.

If we download this data (e.g. through the UI option of Profile -> Download
JSON dump), then it won't have `data.fmt_time` and `metadata.write_fmt_time`
filled in.

If we load this data directly into the timeseries, future pipeline steps will
break since they expect `data.fmt_time` entries.

Let us instead load unprocessed entries into the usercache. This will also
allow us to test the entire pipeline, including the `move to long-term` stage
and the associated formatting, as part of the load process.