Remove implementation of category_of_business
Our current plan is to first have the business id and then look up the
categories from it. That is a simple API call, and does not need this much
testing. We do not currently plan to use `category_from_address` and there is
no proof that it actually worked anyway.