Read the OTP server URL from the OTP_SERVER environment variable
Still debating a bit about environment variable v/s conf file.
But environment variable is clearly easier for docker to deal with and the conf
file may be overkill given that there is only one variable.

We pass in the server URL to the OTP class, and then use the environment
variables in the calling functions for now.

Also remove the `planner.json` file since it is no longer used