Support both "times" and "timestamps"
The `tracejson` format has some open questions around how to represent start
and end times. Should they also be milliseconds? What should the labels be, etc?
In order to avoid dealing with those issues right now, we return both "times"
and "timestamps". The "times" retain backward compatibility with existing code;
the "timestamps" ensure that it is valid tracejson.

We will revisit this after the paper on mobility standards:
https://github.com/e-mission/e-mission-docs/issues/542#issuecomment-667602989