Replace deprecated `count()` method on the pymongo cursor
MongoDB has deprecated `count()` on cursors and has no plans of ever restoring it
https://jira.mongodb.org/browse/PYTHON-1724

We used `Collection.count()` and `Cursor.count()` extensively.
This commit replaces all of those with the new API calls.

Concretely, the find and replace regex was:
- Collection.`find\(.*\).count()` ➡️  `count_documents\1`
- Collection.`find().count()` ➡️  `estimated_document_count`
- Collection.`count()` ➡️  `estimated_document_count()`

In a couple of files, we returned a cursor from a function, checked the count
and then used values from the cursor. I had to replace these with a separate
call to get the count.

These were in:
- `emission/storage/timeseries/builtin_timeseries.py`
- `emission/storage/decorations/section_queries.py`
- `emission/storage/decorations/stop_queries.py`

I also had to change `$where` ➡️  `$expr` in
`emission/analysis/modelling/tour_model/prior_unused/exploratory_scripts/explore_smoothing_trajectories.py`