Fix the expected number of sections and stops
In this test, we count the number of stops and sections for a trip from Mountain View to Berkeley.
On the master branch, there are 11 sections and 10 stops
https://github.com/e-mission/e-mission-server/blob/e218b59a80edcd8afea3063b961d21d387b578d9/emission/tests/analysisTests/intakeTests/TestSectionSegmentation.py

Because we have improved the segmentation, it is now down to 8 sections and 7 stops.
The relationship between stops and sections is consistent, (sections = stops +
1), and the number of sections is lower with our improvements

The current list of sections is

2015-11-06T06:52:41.559255-08:00 2015-11-06T06:53:14.724869-08:00 MotionTypes.WALKING
2015-11-06T06:53:21.165464-08:00 2015-11-06T06:55:11.119814-08:00 MotionTypes.BICYCLING
2015-11-06T06:57:03.128348-08:00 2015-11-06T07:27:58.558022-08:00 MotionTypes.IN_VEHICLE
2015-11-06T07:34:35.440815-08:00 2015-11-06T08:25:06.814844-08:00 MotionTypes.IN_VEHICLE
2015-11-06T08:26:34.972098-08:00 2015-11-06T08:27:32.895544-08:00 MotionTypes.WALKING
2015-11-06T08:29:42.622109-08:00 2015-11-06T08:30:53.610086-08:00 MotionTypes.BICYCLING
2015-11-06T08:31:32.324552-08:00 2015-11-06T08:33:03.106796-08:00 MotionTypes.BICYCLING
2015-11-06T08:34:47.945336-08:00 2015-11-06T08:43:37.125328-08:00 MotionTypes.WALKING

It would probably have been better to combine the two `BICYCLING` entries but
this is better than what we had before, so we are updating it.