Ensure that we register the user email on more complex tests as well
This is a follow-on to
https://github.com/e-mission/e-mission-server/commit/0fd0fd81be75c6e30cf7f36ebc3a1d859bfa1139

For the more complex tests, we were calling `etc.setupRealExampleWithEntries` directly
which didn't create a new UUID. So the test created the UUID using
`self.testUUID = uuid.uuid4()` and our comparison registration didn't work.

So we refactor the UUID creation code into a separate function and call that
from the complex tests instead.