Support fcm tokens for iOS phones as well
By default, the FCM plugin generates FCM tokens for both android and iOS.
I don't like this on principle because it makes it harder to migrate to a
different push service later.

So we typically configure the app to generate APNS tokens and map on the server
side. But that now requires editing GoogleServices.plist (FCM is the default)
and it is easy to forget to do that.

However, others may also not have this concern, and it does simplify the server
side logic significantly to not have to do the mapping.

So I now support FCM on both platforms, and add a config option to specify
whether we are using APNS or FCM for iOS.

If we are using FCM, we can skip the entire mapping logic