Calculate the overall score
We define this is the multiple of the confirmedPct and the validPct.
This is inspired by the conditional probability.

If all trips were using eBikes, then the probability of getting a usable trip
is the probability of getting a confirmed trip multiplied by the probability of
the confirmation being valid.

We sort on the overallScore and return it from the API