Not sure if this is necessary
But since we now have some participants who are not in the leaderboard,
we apparently get None for their user tier, which breaks the lookup here.
Seems like a good check overall even if I am not sure why it occured when I was
trying to login as a user who was in fact on the leaderboard.