Return None for the range if there are no analysis results
This allows us to not go into an infinite scroll in case the user doesn't have
any analysis results.

Without this fix, we returned (-1, -1), so the phone code assumed that was
valid and created the timestamp range by subtracting a week from it. So we kept
going into negative timestamp ranges and got into an infinite loop.

This returns None, so the `if (pipeline_range.end_ts)` check actually works