Improve the implementation of the accuracy filtering step
This is an initial fix for
https://github.com/e-mission/e-mission-docs/issues/637

Concretely, instead of using our home-grown versions of `check_prior_duplicate` and
`check_existing_filtered_location`, we use built-in pandas methods.

For `check_prior_duplicate`, we `drop_duplicates`
For `check_existing_filtered_location`, we can merge the existing locations with an inner join

Then, instead of making n queries to find the entries, set the new metadata key, and save them,
we make one query to get all the `background/location` entries and just iterate
through the in-memory entries to save the new `background/filtered_location`
entries

+ add a new test step `checkSuccessfulRun` to ensure that the step ran
correctly since otherwise. Otherwise the `testEmptyCall` in particular, can
generate an error but not fail. The code catches errors and marks the step as failed.

Testing done:
- Existing tests pass