Summary of changes made: The file conf/log/export.conf.sample has been commited to the branch. I have removed the #skip entry with mixed time and distance... code from both export_stage.py and intake_stage.py as we discussed it is outdated. I have changed the class name in the testing infastructure to match the file name. I have established a common location for the export data within emission/archived.
Testing Done:
The unit tests for the export code can now be found under emission/tests/exportTests/TestExportPipelineConfirmedTripsandBackground.py.
The testing conpares the raw data with the exported data, making sure that we have equal total inputs by comparing the count of data entries with the key for background location (this is present in both raw and exported). It also compares to confirm that we have confirmed trips for the exported and not within the raw.