Drastically rewrite the pipeline code to deal properly with the early return
If you are resetting to a point where the full pipeline has not run, but some
steps have, you still need to reset. This is particularly important when if the
pipeline crashes part-way through a stage and some of the output objects are
persisted while others are not.

This fixes https://github.com/e-mission/e-mission-docs/issues/645#issuecomment-876923027
+ add a new, successful unit test