Improve the implementation of `get_all_uuids()`
The previous implementation was in a situation where we didn't know for sure
whether or not there would be UUID entries for each user. But that has been
true by default for years now.

If we ever got into a situation in which we didn't have the uuid_db, we would
just create the missing entries.

Switching to the UUID DB avoids weird UUIDs like `None` and `null`, and also
is significantly faster.

```
%timeit edb.get_uuid_db().distinct("uuid")
1.29 ms ± 12 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```

```
%timeit edb.get_timeseries_db().distinct("user_id")
3.68 s ± 105 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```