New demo placeholder + label specific reset
We have a cold start problem with demos - we want to be able to showcase the
labeling, but it won't work until we have sufficient labels.

This commit introduces a new placeholder that will help with demos. It pulls
the unique set of existing labels and randomly returns one of them 9/10th of
the time. The remaining 1/10th of the time, it returns no inference.

This lets us prototype our goal, where most of the trips are labeled, and the
user only occasionally has to input labels.

However, if the user does have sufficiently good labels, we do want to use
that.  So we use our built-in ensemble method to generate two labels, pick the
"real", non-placeholder if if exists, and fallback to the placeholder if it
doesn't.

Note that if we enhanced the placeholder to use the actual distribution of the
existing labels, this could arguably be a very simple randomized algorithm.

+ added a super simple script to reset only the label inference stage and the
only the confirmed trip stage so that we can experiment with this without
having to re-run the complete pipeline

+ added a super simple script to print out label stats - how many of each kind of trip, to debug pipeline operation

Testing done:
- Reset the inference pipeline
- Ran the intake pipeline

Placeholder was invoked

```
2021-07-20 10:54:54,758:DEBUG:4457897408:In placeholder_predictor_demo: ound 85 for user 1d292b85-c549-409a-a10d-746e957582a0, returning value {'mode_confirm': 'shared_ride', 'purpose_confirm': 'meal', 'replaced_mode': 'same_mode'}
2021-07-20 10:54:54,866:DEBUG:4457897408:In placeholder_predictor_demo: ound 85 for user 1d292b85-c549-409a-a10d-746e957582a0, returning value {'mode_confirm': 'walk', 'purpose_confirm': 'exercise', 'replaced_mode': 'bus'}
2021-07-20 10:54:54,972:DEBUG:4457897408:In placeholder_predictor_demo: ound 85 for user 1d292b85-c549-409a-a10d-746e957582a0, returning value {'mode_confirm': 'skateboard', 'purpose_confirm': 'exercise', 'replaced_mode': 'same_mode'}
```

Final result was a combination of real and placeholder

```
2021-07-20 10:54:55,182:DEBUG:4457897408:Found real prediction Labelprediction({'trip_id': ObjectId('60f70def49a6d6de8e8ef803'), 'algorithm_id': 4, 'prediction': [{'labels': {'mode_confirm': 'not_a_trip', 'purpose_confirm': 'home', 'replaced_mode': 'not_a_trip'}, 'p': 1.0}], 'start_ts': 1612036947.149, 'end_ts': 1612037455.46}), using that preferentially
2021-07-20 10:54:55,286:DEBUG:4457897408:No real prediction found, using placeholder prediction Labelprediction({'trip_id': ObjectId('60f70def49a6d6de8e8ef807'), 'algorithm_id': 5, 'prediction': [{'labels': {'mode_confirm': 'walk', 'replaced_mode': 'same_mode', 'purpose_confirm': 'meal'}, 'p': 0.9517281073260278}], 'start_ts': 1612132424.3783095, 'end_ts': 1612133524.871})
```

- Reset confirmed trips

```
all inferred trips 215
all confirmed trips 0
confirmed trips with inferred labels 0
confirmed trips without inferred labels 0
confirmed trips with expectation 0
```

- Reran the pipeline

```
all confirmed trips 215
confirmed trips with inferred labels 215
confirmed trips without inferred labels 0
confirmed trips with expectation 215
```