Add simple script to federate multiple datasets and save them to a csv
We were going to do the analysis here too, but loading the databases at the
same time takes too much memory and disk for regular docker settings. I had to
bump up the settings to get everything to work.

This two stage process helps others who may not have as robust computers to
also experiment with the combined dataset