Use the trip start and end location directly
Instead of looking up the place and getting it instead.

This has two advantages over the current implementation:
1. We don't have to make 2 separate database calls for each trip
    Note that we compute an nxn distance matrix, so this is likely to be a
    substantial savings
2. We can pass in a in-memory trip list. That makes it easier to write unit
    tests, and to use alternate load methods (e.g. for working with federated data
    https://github.com/e-mission/e-mission-eval-private-data/pull/28/commits/952c476dec745eb66477fed8d699f1bab85d84c0

@corinne-hcr reported that the place location and the trip start/loc locations
are not identical. We don't have unit tests to verify this (alas!) but the top
level results are not changed significantly.

So the ROI seems high enough; we are going ahead with this change.
https://github.com/e-mission/e-mission-server/pull/826#issuecomment-885714205