Create a function to expand the userinput into separate dataframe columns
If the labels are simple KV pairs, the result columns are scalar.
Currently duplicated from:
https://github.com/e-mission/em-public-dashboard/blob/main/viz_scripts/scaffolding.py
but without the ipython dependencies and the assumption that there will always
be three entries in the userinput. We removed the assert to handle this assumption.

TODO: Consider replacing with pandas.io.json.json_normalize