Support multiple usage patterns + return labels
This commit has two main components.

The simpler component is to simply pull out the trip filteration code into a
separate function that we can invoke externally instead of embedding it into
the init. This allows us to experiment with the clusters both with and without
filtering. Note that, if we want to be consistent with sklearn, we should
convert this to a set of parameters instead.

The second component is a function to generate "label" for all the trips that
are passed in.

This is pulled out into a separate function because of the tricksiness of
filtering followed by binning, in which the indices in the bins are from the
filtered trips and not the complete set of trips passed in.

So if we passed in 20 trips (indices = `range(20`)
and the following trips were filtered out:
`[0, 5, 10, 15]`
the bins would be

```
[[ 0  1  2  3]
 [ 4  5  6  7]
 [ 8  9 10 11]
 [12 13 14 15]]
```

However, using those indices directly will return labels
`[0,0,0,0,1,1,1,1...]`. So trip 5 would get label of 1 although it was actually
filtered out.

To fix this, we store trip arrays of the various stages of the algorithm, and,
when appropriate, map from bin indices -> all trip indices before setting the
labels.

Testing done:
Added tests for generating labels for various combinations of filtering, binning and cutoffs.
All tests pass.