Finish the multi-dataset visualizations, at least for now
Major changes were to integrate the functions from the "unrolled" exploration
code into this notebook so that we could generalize the "oursim" results across
multiple datasets as well.

It also appeared that we were building all the models **twice** for reasons
that I don't understand; unified it to build the models once. Also pulled out
the summary statistics code into a separate cell, so we can experiment with it
without having to re-run the models, which take significant time.

Added several visualizations to visualize the findings, which are:
- oursim and DBSCAN have similar request pct
- oursim has higher homogeniety score than DBSCAN
- we have lower noise and tighter models if we have 100+ labeled trips

Visualizations at:
https://github.com/e-mission/e-mission-eval-private-data/pull/28#issuecomment-886225003

@corinne-hcr, at this point, I'm going to move on to actually implementing this
in the pipeline so we can get some results before the next meeting on Tuesday.