Check in initial copies of the current modeling code
We copy them instead of modifying the existing models so that we can run them
side by side for a comparison.

TODO: Unify with parameters (at least for the auxiliary files) instead of
making copies.

This is currently unchanged so that we can have a clean diff when we do apply
the changes.