Add a very high-level technical overview + get ready to make the columns side-by-side
Add a very high-level technical overview - targeted at a program admin level -
along with a brief explanation of the "intensive" and "relaxed" modes and why
they are necessary.

+ get ready to put the columns side by side by adding `col` `<div`s within the
`row`. However, to actually make them appear side by side, you need an
additional `col` class.

e.g.

```
<div class="col-md-33 mx-auto" style="text-align: left;">
```

to

```
<div class="col col-md-33 mx-auto" style="text-align: left;">
```

Implementing that, however, makes the images appear side by side, and the
layout is not really responsive until the width is unreasonably small. It would
not convert to a vertical layout on an iPhone6, for example.

So we have backed out that change until we have a chance to look at settings
for this CSS template.