Address the case in which there are no labeled trips in a time grouping
In e2be8bf19390294e19ca7cb43c19e27509bfe804, we addressed the case in which
there was a mixture of labeled and unlabeled trips, by ensuring that the NaN
from the unlabled trips was renamed to "unknown".

However, if there are *no* labeled trips in a time grouping, then the
`expand_userinputs` does not add any columns to the dataframe, so there are no
NaN values to replace.

We fix this by checking to see if the column has been added, and if not, then
adding a column with all NaN. This will then be replaced by "unknown" per
e2be8bf19390294e19ca7cb43c19e27509bfe804, and everything will work

Testing done:
- Added two new tests where we loaded data without user inputs for a different
  month (similar to e2be8bf19390294e19ca7cb43c19e27509bfe804), but then grouped
  by day or month. The last day or month then had all unlabeled trips.
- Created a new `TestMetricsConfirmedTripsPandas` test to test the hardcoded
  pandas behavior since we don't need to load data and run the pipeline in that
  case.
    - Moved the existing `testPandasNaNHandlingAndWorkaround` from there and
      added a new test for this new behavior.