Add support for resetting the labeling only for individual users
Allow users to specify the userid or the user email, or invoke the reset for
all users. We create a base query that has the user in it, and then tack on the
specific key. We use `dict(a, **b)` instead of update to keep the mongodb
commands to a single line each.