Replace median_speed by mean_speed
Keep a redirect from median_speed to mean_speed for backward compatibility
We should remove this redirect in ~ 1 year.

We primarily use the median speed for the calorie calculation, but:
- it is not clear that we need the median speed
- it is harder to calculate for trip-based entities
https://github.com/e-mission/e-mission-docs/issues/688#issuecomment-998466419

So we switch to the mean speed instead. This is currently the grand mean (sum
of all distances / sum of all durations), but we can fairly easily change it to
mean of means if we want.

+ remove _get_speeds_for_trip
+ fix answers in `testAllTimestampMetrics`
+ change `testAllTimestampMetrics` to retrieve both mean and median speeds for now

Testing done:
`testAllTimestampMetrics` passes