Create a new function to expand final labels
This will expand the user labels if they exist, and expand the highest
probability inferred label if we did not prompt the user.

Basically, this is now consistent with
https://github.com/e-mission/e-mission-docs/issues/688#issuecomment-1000981037

There were some minor trickinesses working with this data structure and pandas.
- We had to expand the expectation field to filter by to_label and then
- We had to create a separate dataframe for the inferred labels and the apply a
  custom filter to them to pick the one with the max probability
- We had to ensure that rows didn't appear twice - i.e. once for user inputs
  and once for inferred labels. This is true even if the user_inputs were not
  filled out; without changes, we would end up with both N/A and the inferred
  label for the same row.

And we had to keep the indices constant throughout these changes.
Added multiple tests, both of the individual pandas steps and the complete
function to test this fairly complex piece of code.

Testing done:
- Ran the newly added tests

Next steps:
- Change the metrics code and the leaderboard to use the new functions