Remove the webapp from the server
As part of modularizing the platform, let's delete the webapp from the server code
We should create a separate webapp microservice, presumably using node.js so
that we can iterate on it separately.

This should also significantly lower the space taken by the e-mission-server
repo, simplify its dependencies, and stop generating dependabot alerts.