Fallback to stripping out study-specific prefix if the user account is not found
This is a backwards compat fix for the issue that we prefix the opcode with the
program name for programs but not for autogenerated study tokens

We will fix this on the phone side, but will need a backwards compat plan for
existing users until their app has been upgraded.

The plan is to keep the server tokens without the study prefix (for now) and to
strip out the study prefix if needed to find the matching token

And once all the phones have been upgraded, we can upgrade the tokens on the
server and be all done.
https://github.com/e-mission/e-mission-docs/issues/732#issuecomment-1179828277

Testing so far:
- Logged in with a valid token

```
2022-07-10 17:37:53,019:DEBUG:123145553195008:Using the skip method to verify id token user-0 of length 6
2022-07-10 17:37:53,024:DEBUG:123145553195008:retUUID = ...
2022-07-10 17:37:53,030:DEBUG:123145553195008:Updating user ...
```

- Logged in with an invalid token

```
2022-07-10 17:47:44,355:DEBUG:123145542684672:retUUID = None
2022-07-10 17:47:44,355:DEBUG:123145542684672:token ... is valid, but no account found. stripping out study name and retrying...

abort(400, "Invalid study format %s, does not end with -openpath" % first_domain)
```

Need to test further on a server with a valid domain name.