Store the model in a new model_db instead of in the analysis_db
Unlike the user inputs, which are read-only from an analysis perspective, or
analysis results, which are write once (except for confirmed trips), the models
are updated frequently, every time they are re-calibrated.

So we create a new interface to match this new access pattern.
Unlike the timeseries, which focuses on reading data for a particular range,
the new interface focuses on updating models and retrieving the most recent model

We still need to write code to purge obsolete models so that the database does
not grow unboundedly.

Changes:
- new database
- new abstract interface in `emission/storage/modifiable`
- new concrete implmentation in `emission/storage/modifiable`
- change the `model_storage` to use the new interface instead of the timeseries
    - bonus, don't need to use the concrete implementation any more since this
      implementation is all about the most recent version
- change the test to delete the model db in addition to the analysis DB
    - we need to now delete both since the trips are in the analysis DB while
      the models are in the separate model DB