Add the FUZZ while resetting the pipeline for the label assist functionality
Without this FUZZ, if we reset the pipeline to time t1, then the trip that ends
at t1 will match the `data.end_ts >= previous end` fix, so will be processed
multiple times.

However, we do not want to add the FUZZ for the previous trips (in particular,
for CLEAN_AND_RESAMPLE) due to
https://github.com/e-mission/e-mission-docs/issues/191#issuecomment-312917119

So we only apply the fuzz factor to trip segmentation and the label assist
stages.

We don't apply it to the other stages.

With this change, the automated tests pass.

This fixes: https://github.com/e-mission/e-mission-docs/issues/806