Create a script to automatically reset the pipeline to the correct value for each user
This replaces the current, cumbersome method in which I get a list of users
with an invalid pipeline state, and then reset them manually.

This automates that manual algorithm, while also making it more sophisticated.

The algorithm essentially finds all invalid states for a current user, groups
them, and then resets to the oldest timestamp. This will ensure that we don't
skip some data by moving the pointer ahead to a time that is not yet
processed.

This fixes https://github.com/e-mission/e-mission-docs/issues/805