Wrap a mongodb cursor in a list before converting it to a dataframe
Using `pd.json_normalize` directly on a cursor seems to skip the first line.
Wrapping the cursor in a list seems to fix it.
Not sure if this is expected behavior or whether it has been fixed now in pandas

This change:
- adds a new test that explicitly tests this functionality - the test passes if
  the first line is missing in the cursor but not in the list. This does not
  call any functions in the reset.
- adds a new check to the existing mock test by adding a new user
  (`no_missing`) to the front of the test data and ensuring that the user is
  also reset correctly. Without this change, the first entry had multiple
  repetitions, so even if the first entry was skipped, the user would still be reset
- fixes the issue in the reset code

Testing done:
- tests pass

This re-fixes https://github.com/e-mission/e-mission-docs/issues/805