Ensure that back to back reset runs work
Historically, when we unset the entries for the last cleaned place during
a reset, we set the raw place array to empty. if the next run of the
pipeline also fails, then the last cleaned place will not have any raw
places associated with it. In that case, we find the matching raw place
and fill it in

We deal with this by finding the corresponding ending raw trip and finding the
end raw place so that we can link them together

I thought that we might want to save the raw place as well, but I see that
we have a comment in the code as below so we need to experiment with that later.
And figure out how the squished places should work going forward.

> Note that we need to reset the raw_place array
> since it will be repopulated with new squished places
> when the timeline after the _entry_ to this place is reconstructed

The issue is documented for long-term investigation in
https://github.com/e-mission/e-mission-docs/issues/808#issuecomment-1272464113

Testing done:
Added new test case `testResetTwiceHack`
without this fix, the second reset fails with the same error as encountered in the real world

```
ERROR:root:list index out of range
Traceback (most recent call last):
  File "/Users/kshankar/e-mission/nrel-db-connect/emission/pipeline/reset.py", line 346, in auto_reset
    reset_user_to_ts(invalid_state['user_id'], invalid_state['reset_ts'], dry_run)
  File "/Users/kshankar/e-mission/nrel-db-connect/emission/pipeline/reset.py", line 73, in reset_user_to_ts
    reset_ts, last_cleaned_place, last_raw_place = get_reset_ts(user_id, last_cleaned_place, is_dry_run)
  File "/Users/kshankar/e-mission/nrel-db-connect/emission/pipeline/reset.py", line 90, in get_reset_ts
    last_raw_place_id = last_cleaned_place["data"]["raw_places"][-1]
IndexError: list index out of range
```

With this fix, the test passes