insert.py: added default case and made args mutex
Created a case for when no options are selected and forced options to be mutually exclusive. Not strictly necessary for options to be mutex, but legitimate use cases seems unlikely and having them all separate eliminates the chance of unexpected behavior.

Added new unit tests for updated functionality.

Changed out os() for subprocess.run()