Add functionality to ignore duplicate entries on load
Since we re-generate the files to fix the prior errors with missing entries, we
will have duplicate entries in the dump. We need to be able to restore from the
dump, and end up inserting only the missing entries.

We support this use case by adding the capability to continue on error to the
load script and plumbing that through the code flows, namely the cache_series
and the insertion of the pipeline states. In both cases, if we get a duplicate
key error E1100, we continue instead of failing.

Testing done:
- loaded data, no errors