Handle the matching of user inputs that came in before the confirmed trip was created
In https://github.com/e-mission/e-mission-server/pull/895/commits/21290759cd1a1b3dfcd318ee56e2ac08258c16e4
we added support for matching inputs that came in after the trip was created

However, it is also possible that the input came in before the trip was processed.
In that case, we need to find all input matches for the trip and fill it in when the trip is created

This is currently implemented in `final_candidate` for the single most recent match
We add a new function `get_not_deleted_candidates` that is similar, and write unit tests for both of them.

We then add a simple wrapper `get_additions_for_trip_object`, similar to `get_user_input_for_trip_object`
and use it to fill in the `confirmed_trip_dict["data"]["trip_addition"]` while creating confirmed trips

Testing done:
- unit tests pass