Added unit test for `None` backup algo + unify algo outputs
Added a new unit test for the case of `backup_algo == None`, which should
return the original algo results.

While testing, found that the ZigZag algo returns a pandas Series,
while the Posdap algo returns a numpy array, which means that combining them
could be problematic

Changed ZigZag to also return a numpy array to unify the implementations.
Testing done:
- All tests now pass