:art: Return and record the selected algo correctly
Before this change, we only used one algorithm, so we hardcoded it into the
result. However, we can now use either the main algorithm or the backup
algorithm. So we return the algo also from `get_points_to_filter` and attribute
it correctly.

`get_points_to_filter` is used only in `location_smoothing` and in the tests.
So also fix the tests to read both values and check the sel algo in each case

Testing done: tests pass