Fix regression caused by moving the second round checking out
When we moved the second round checks to the calling function in
cebb81fac17d1e4684d3b76754ecabaa49af1a57
we caused a very subtle regression

The filtering code had an early return if there were no jumps detected.
So in that case, we would not try the second round of checks, or attempt to
filter again.

However, when we moved the second round checking to the outer function, we
called the second round anyway even if the first round didn't detect any jumps
And in this one case, we actually found an outlier in the second round, which
caused the test to fail.

Fixed by checking to see if there were no outliers in the first round and
skipping the second round check in that case.

Everything in the `else` for the
`if outlier_arr[0].shape[0] == 0:` is unchanged, just moved in a bit, not changed.

The check for the length was unexpectedly complicated and took many hours to
debug, so I added it as a simple use case.

Note also that it is not clear if this is the correct long-term approach.
If there were no jumps, then why did using the backup change anything?
Maybe we should always use the backup.

But changing this to avoid the regression for now; will look at this the next
time we look at smoothing

Testing done:
- `TestPipelineRealData.testIosJumpsAndUntrackedSquishing` passes
- `TestLocationSmoothing` passes