:white_check_mark: Fix regression in `TestExpectationPipeline` + add tests to automated CI
`TestExpectationPipeline` assumes that there are `mode_confirm` and
`purpose_confirm` objects, because they will match and have a high confidence.
If we add a new key `user_input`, then there are no matches for it, so we end
up with the `-1` option (see details below).

This is a regression caused by
https://github.com/e-mission/e-mission-server/commit/ae325ecbdddb9c2c6c90b1f874d63983a21069eb

We fix this by changing the config to only include mode and purpose confirm
during setup using the built-in `override_keys` and by restoring the config at
the end.

Also add some code to help debug this (currently commented out)

Testing done:
- Tests pass

Expectation test failure:
- test is failing with trip 795
    the date is `arrow.get("2021-06-08T20:00:00.000", tzinfo=tz),`
    we expect `795: {"type": "none"},`
    we get `{'type': 'randomDays', 'value': 2}`

The only config which has `randomDays` is the relaxed configuration, and we
should only see that for the `-1` trigger. In our case, though, we have a label

```
[{'labels': {'mode_confirm': 'bike', 'purpose_confirm': 'work'}, 'p': 0.96}, {'labels': {'mode_confirm': 'walk', 'purpose_confirm': 'shopping'}, 'p': 0.04}]
```

we also have

```
'expectation': {'to_label': True}, 'confidence_threshold': 0.55}
```

which means that it should have returned

```
                {
                    "trigger": 1,
                    "expect": {
                        "type": "none"
                    }
                }
```

why did that not happen?

mode is

```
{'label': 'relaxed', 'confidenceThreshold': 0.55, 'rules': [{'trigger': -1, 'expect': {'type': 'randomDays', 'value': 2}, 'notify': {'type': 'dayEnd'}}, {'trigger': 1, 'expect': {'type': 'none'}}, {'trigger': 0.95, 'expect': {'type': 'randomFraction', 'value': 0.05}, 'notify': {'type': 'dayEnd'}}, {'trigger': 0.75, 'expect': {'type': 'randomDays', 'value': 2}, 'notify': {'type': 'dayEnd'}}]}
```

But the userinput key list (`for input in _keylist:`)
https://github.com/e-mission/e-mission-server/blob/b7749d0a33f579eab8362416c32d2dcebd4d0b0c/emission/analysis/configs/expectation_notification_config.py#L106
populated from
https://github.com/e-mission/e-mission-server/blob/b7749d0a33f579eab8362416c32d2dcebd4d0b0c/emission/analysis/configs/expectation_notification_config.py#L30

also includes `userinput`, which does not have any matches, so it matches the `-1` trigger