Using uuidRepresentation as 'pythonLegacy'
We are using 'python_legacy' instead of 'standard' UUID representation here because standard UUID representation should be used by new applications or applications that are encoding and/or decoding UUIDs in MongoDB for the first time. Python_legacy should be used by already existing applications using python drivers to encode and/or decode.

Ours is a python application that has already been storing data in databases for many months. It is NOT "encoding or decoding for the first time". If we used standard, although the unit tests might run, it would break in production.

Contains fix for the issue:
https://github.com/e-mission/e-mission-docs/issues/856#issuecomment-1431879445